10 - Users can share something
- As title says, users can share passwords.
- Operators aren't still possible.

+++ PasswordManager/app/src/main/java/it/progmob/passwordmanager/SettingsFragment.kt
@@ -1,14 +1,18 @@
 package it.progmob.passwordmanager
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import androidx.navigation.Navigation
+import androidx.navigation.fragment.findNavController
 import com.firebase.ui.auth.AuthUI
 import com.google.firebase.Firebase
 import com.google.firebase.firestore.firestore
@@ -35,28 +39,51 @@ class SettingsFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.resetButton.setOnClickListener {
-            val db = Firebase.firestore
+            val dialogBuilder = AlertDialog.Builder(requireContext())
+            val viewInflated: View = LayoutInflater.from(requireContext()).inflate(R.layout.confirm_layout, null, false)
 
-            db.collection("Passwords").get()
-                .addOnSuccessListener { documents ->
-                    for (document in documents) {
-                        document.reference.delete()
-                    }
-                };
-            db.collection("Pins").get()
-                .addOnSuccessListener { documents ->
-                    for (document in documents) {
-                        document.reference.delete()
-                    }
-                };
-            db.collection("CreditCards").get()
-                .addOnSuccessListener { documents ->
-                    for (document in documents) {
-                        document.reference.delete()
-                    }
-                };
+            dialogBuilder.setView(viewInflated)
+            dialogBuilder.setTitle("You're going to reset everything, are you sure?")
+            dialogBuilder.setCancelable(false)
 
-            viewModel.reset()
+            // Buttons
+            dialogBuilder.setPositiveButton("Submit") { _, _ -> }
+            dialogBuilder.setNegativeButton("Cancel") { dialog, _ ->
+                dialog.cancel()
+            }
+
+            val alertDialog = dialogBuilder.create()
+            alertDialog.show()
+            alertDialog.getButton(AlertDialog.BUTTON_POSITIVE)?.setOnClickListener {
+                val db = Firebase.firestore
+
+                val userRef = viewModel.userID?.let { it1 -> db.collection("users").document(it1) }
+
+                userRef?.collection("Passwords")?.get()
+                    ?.addOnSuccessListener { documents ->
+                        for (document in documents) {
+                            document.reference.delete()
+                        }
+                    };
+                userRef?.collection("Pins")?.get()
+                    ?.addOnSuccessListener { documents ->
+                        for (document in documents) {
+                            document.reference.delete()
+                        }
+                    };
+                userRef?.collection("CreditCards")?.get()
+                    ?.addOnSuccessListener { documents ->
+                        for (document in documents) {
+                            document.reference.delete()
+                        }
+                    };
+
+                viewModel.reset()
+
+                alertDialog.dismiss()
+
+                Toast.makeText(requireContext(), "Everything has been deleted correctly.", Toast.LENGTH_LONG).show()
+            }
         }
 
         binding.logoutButton.setOnClickListener {
